fix infinite recursion
